Resolve import cycle by extracting registration models into registrationmodel package
