+++ pkg/registrationmodel/registration.go
@@ -1,4 +1,4 @@
-package registration
+package registrationmodel
 
 import "time"
 
@@ -25,13 +25,6 @@ type Registrant struct {
 	Name   string `bson:"name" json:"name" binding:"required"`
 	Email  string `bson:"email" json:"email" binding:"required,email"`
 	UserID string `bson:"user_id" json:"user_id"`
-}
-
-// SubmitRegistrationRequest represents the incoming API payload from the frontend when a user registers for an event
-type SubmitRegistrationRequest struct {
-	Registrant  Registrant     `bson:"registrant" json:"registrant" binding:"required"`
-	Answers     map[string]any `bson:"answers" json:"answers" binding:"required"`
-	SubmittedAt string         `bson:"submitted_at,omitempty" json:"submitted_at,omitempty"`
 }
 
 // RegistrationRequest is the source-of-truth document representing a user's registration lifecycle (pending to accepted/rejected)
